+++ src/primitives/stealth.cpp
@@ -58,7 +58,7 @@ stealth::stealth(const stealth_prefix& prefix, const ec_public& scan_key,
     const std::vector<ec_public>& spend_keys, uint8_t signatures, bool testnet)
 {
     // Normalize signatures between 1 and spend_keys.size().
-    const auto size = if_else(spend_keys.empty(), 1, spend_keys.size());
+    const auto size = if_else(spend_keys.empty(), (size_t)1, spend_keys.size());
     auto sigs = if_else(signatures == 0, size, signatures);
     sigs = if_else(sigs > size, size, sigs);
